fix(domain): remove unused imports to clean up code refactor(domain): relocate DecodeJSON function to validators.go for better organization style: remove unused time import from generate.go
